Fixed databuffer.c crashing when storing values of int type

+++ include/databuffer.h
@@ -6,23 +6,23 @@
 
 typedef struct data_buffer_f
 {
-    float* element;
-    size_t count;
-    size_t size;
+    DWORD* element;
+    DWORD count;
+    DWORD size;
 }data_buffer_f_t;
 
 typedef struct data_buffer_i
 {
-    int64_t* element;
-    size_t count;
-    size_t size;
+    DWORD* element;
+    DWORD count;
+    DWORD size;
 }data_buffer_i_t;
 
-data_buffer_f_t* data_buffer_new_f(size_t);
+data_buffer_f_t* data_buffer_new_f(DWORD);
 void data_buffer_destroy_f(data_buffer_f_t*);
 int data_buffer_add_element_f(data_buffer_f_t*, float);
 
-data_buffer_i_t* data_buffer_new_i(size_t);
+data_buffer_i_t* data_buffer_new_i(DWORD);
 void data_buffer_destroy_i(data_buffer_i_t*);
 int data_buffer_add_element_i(data_buffer_i_t*, int32_t);
